shipment/requirement.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.15.1
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
isodate==0.6.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
looseversion==1.3.0
lxml==5.2.2
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.0.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
prov==2.0.1
psycopg2==2.9.9
pycparser==2.22
pydot==2.0.0
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tabula-py==2.9.3
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2